packages/backend/requirements.txt
python-dotenv
eth-brownie